// src/data/experimental.generated.ts
// Generated by pnpm scrape:experimental. Do not edit manually.

import type { PartialWeaponCatalog } from '../data';

export const experimentalWeapons: PartialWeaponCatalog = {
  "Flanker": {
    "Primary": [
      {
        "name": "All Sides",
        "image": "images/experimental-generated/all-sides-9d4a9d3e7e.webp"
      }
    ]
  },
  "Trooper": {
    "Secondary": [
      {
        "name": "Ranger's Rifle",
        "image": "images/experimental-generated/ranger-s-rifle-e6e59fab83.webp"
      }
    ]
  },
  "Arsonist": {
    "Primary": [
      {
        "name": "Napalmer",
        "image": "images/experimental-generated/napalmer-d37053fa01.webp"
      },
      {
        "name": "Storage Stuffer",
        "image": "images/experimental-generated/storage-stuffer-0b87f01920.webp"
      },
      {
        "name": "Immolator",
        "image": "images/experimental-generated/immolator-cdc6cef6ad.webp"
      }
    ]
  },
  "Annihilator": {
    "Primary": [
      {
        "name": "Aggressor",
        "image": "images/experimental-generated/aggressor-e36bef6d34.webp"
      }
    ]
  },
  "Brute": {
    "Secondary": [
      {
        "name": "Fiber Bar",
        "image": "images/experimental-generated/fiber-bar-16b37f4594.webp"
      },
      {
        "name": "Special Meal",
        "image": "images/experimental-generated/special-meal-c8eeacb9bb.webp"
      }
    ],
    "Melee": [
      {
        "name": "Feedbacker",
        "image": "images/experimental-generated/feedbacker-16282faa74.webp"
      },
      {
        "name": "World Champion",
        "image": "images/experimental-generated/world-champion-10e01e704c.webp"
      }
    ]
  },
  "Mechanic": {
    "Secondary": [
      {
        "name": "Boomwarden",
        "image": "images/experimental-generated/boomwarden-043a60cf39.webp"
      },
      {
        "name": "Mechanical Marvel",
        "image": "images/experimental-generated/mechanical-marvel-8c4946f874.webp"
      },
      {
        "name": "Wrecking Ball",
        "image": "images/experimental-generated/wrecking-ball-b6cc60f888.webp"
      }
    ]
  },
  "Doctor": {
    "Secondary": [
      {
        "name": "Barrett's Privateer",
        "image": "images/experimental-generated/barrett-s-privateer-e7f76ce02c.webp"
      },
      {
        "name": "Neuron Connection",
        "image": "images/experimental-generated/neuron-connection-1309a8155c.webp"
      }
    ]
  },
  "Marksman": {
    "Primary": [
      {
        "name": "Slick Shot",
        "image": "images/experimental-generated/slick-shot-04921ebcc4.webp"
      },
      {
        "name": "Silent Hunter",
        "image": "images/experimental-generated/silent-hunter-98c10616a6.webp"
      }
    ]
  },
  "Agent": {
    "Primary": [
      {
        "name": "Double Agent",
        "image": "images/experimental-generated/double-agent-b6cc60f888.webp"
      },
      {
        "name": "Sinister Suit",
        "image": "images/experimental-generated/sinister-suit-54ffa9a07f.webp"
      }
    ],
    "Melee": [
      {
        "name": "Argento Carro",
        "image": "images/experimental-generated/argento-carro-c2a4e50a91.webp"
      },
      {
        "name": "Trinity",
        "image": "images/experimental-generated/trinity-81e08dd53a.webp"
      },
      {
        "name": "Wraith",
        "image": "images/experimental-generated/wraith-cc92744e90.webp"
      }
    ],
    "PDA": [
      {
        "name": "Afterimage",
        "image": "images/experimental-generated/afterimage-8f7d7ca26b.webp"
      }
    ]
  }
};